Airflow/requirements.txt
requests~=2.32.3
pendulum~=3.0.0
pandas
ijson
findspark
matplotlib
pyspark
python-dotenv
supabase
six
kafka-python-ng
apache-airflow
elasticsearch==7.8.1
elasticsearch-async